Improving slider and buttons and layering
Slider now only transitions when you press the arrow keys or the buttons. It was transitioning during a page resizes and looking really weird.
The left and right arrow now have press detection for the entire height of the page so that it is easier to click on them.
The text is now on the right side of the page like a column in a newspaper.
The bottom-image-slider is now anchored to the bottom of the page so that its navigation buttons are always visible.